+++ lib/arxiv/downloader.rb
@@ -10,6 +10,7 @@ require_relative 'downloader/path'
 require_relative 'downloader/pdf'
 require_relative 'downloader/bibtex'
 require_relative 'downloader/abstract_page'
+require_relative 'downloader/source_archive'
 
 module Arxiv
   module Downloader